requirements.txt
# Core dependencies
streamlit
langgraph
langchain
langchain-openai
langchain-anthropic
langchain-google-genai

# Graph and visualization
networkx
matplotlib
plotly
pyvis

# Image processing
opencv-python
Pillow
scikit-image

# OCR and text extraction
pytesseract
easyocr

# Computer Vision / ML
torch
torchvision
ultralytics  # For YOLO
transformers

# PyFlowsheet integration
# pyflowsheet  # Uncomment if available

# Data processing
numpy
pandas
pydantic

# Utilities
python-dotenv
pyyaml
tqdm

# Visualization
graphviz
seaborn

# Optional: Enhanced features
pdf2image  # For PDF PFDs
cairosvg  # For SVG export